Quick hack to solve the subtitles issue
The subtitles on serialepenet.ro are not all in the same format. Some are SRT, some are TTML,
as far as I found out so far. TTML files need to be converted to SRT, which was implemented in
the first place. I added some code to keep SRT files as they are.